+++ include/sg/option/threadingmanager_omp.hpp
@@ -21,7 +21,7 @@ private:
     SpinLock startup_lock;
     SpinLock lock_workers_initialized;
 
-    size_t start_counter;
+    int start_counter;
     char padding1[Options::CACHE_LINE_SIZE];
     int num_cpus;
